refactor(extracts withSubscription, consolidates tsdoc)
