+++ libs/react/src/lib/react.tsx
@@ -1,24 +1,4 @@
-import React from 'react';
-import { Observable } from 'rxjs';
-
 export { useStore } from './use-store';
 export { useSubscription } from './use-subscription';
+export { withSubscription } from './with-subscription';
 export { store, StoreArg, StoreReturn } from './store';
-
-export function withSubscription<T>(
-  WrappedComponent: React.ComponentType<{
-    next?: T;
-    error: any;
-    complete: boolean;
-  }>,
-  source: Observable<T>
-) {
-  const [next, error, complete] = useSubscription<T>(source);
-  return (
-    <WrappedComponent
-      next={next}
-      error={error}
-      complete={complete}
-    ></WrappedComponent>
-  );
-}
